+++ Bow.h
@@ -1,42 +1,38 @@
 #pragma once
 #include "random.h"
 
-// ·ÉÀ×Ö®ÏÒÕñ
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Thundering_Pulse(Tao chara)
+// é£é›·ä¹‹å¼¦æŒ¯
+auto Thundering_Pulse(derived_from_character auto chara)
 {
-	chara.baseATK += 608;
+	chara.add_baseATK(608);
 	chara.value[CRD] += 0.662;
 	chara.value[ATK] += 0.2;
 	chara.value[bonus] += 0.4;
 	return chara;
 }
 
-// µ¯¹­
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Slingshot(Tao chara)
+// å¼¹å¼“
+auto Slingshot(derived_from_character auto chara)
 {
-	chara.baseATK += 354;
+	chara.add_baseATK(354);
 	chara.value[CR] += 0.312;
 	chara.value[bonus] += 0.6;
 	return chara;
 }
 
-// ¹­²Ø
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Rust(Tao chara, int rank)
+// å¼“è—
+auto Rust(derived_from_character auto chara, int rank)
 {
-	chara.baseATK += 510;
+	chara.add_baseATK(510);
 	chara.value[ATK] += 0.413;
 	chara.value[bonus] += 0.3 + rank * 0.1;
 	return chara;
 }
 
-// °¢ÄªË¹Ö®¹­
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Amos_Bow(Tao chara, int time)
+// é˜¿è«æ–¯ä¹‹å¼“
+auto Amos_Bow(derived_from_character auto chara, int time)
 {
-	chara.baseATK += 608;
+	chara.add_baseATK(608);
 	chara.value[ATK] += 0.496;
 	chara.value[bonus] += 0.12 + time * 0.08;
 	return chara;