+++ Catalyst.h
@@ -1,88 +1,118 @@
 #pragma once
 #include "random.h"
 
-// Ç§Ò¹¸¡ÃÎ
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto A_Thousand_Floating_Dreams(Tao chara, int same)
+// åƒå¤œæµ®æ¢¦
+auto A_Thousand_Floating_Dreams(derived_from_character auto chara, int same)
 {
 	if (!(same >= 0 && same <= 3))
 		throw invalid(same);
-	chara.baseATK += 542;
+	chara.add_baseATK(542);
 	chara.value[EM] += 265;
 	chara.value[EM] += 32 * same;
 	chara.value[bonus] += 0.1 * (3 - same);
 	return chara;
 }
 
-// ËÄ·çÔ­µä
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Lost_Prayer_to_the_Sacred_Winds(Tao chara, int time4)
+// å››é£åŸå…¸
+auto Lost_Prayer_to_the_Sacred_Winds(derived_from_character auto chara, int time4)
 {
 	if (!(time4 >= 0 && time4 <= 4))
 		throw invalid(time4);
-	chara.baseATK += 608;
+	chara.add_baseATK(608);
 	chara.value[CR] += 0.331;
 	chara.value[bonus] += 0.08 * time4;
 	return chara;
 }
 
-// Ó¯ÂúÖ®Êµ
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Fruit_of_Fulfillment(Tao chara)
+// ç›ˆæ»¡ä¹‹å®
+auto Fruit_of_Fulfillment(derived_from_character auto chara)
 {
-	chara.baseATK += 510;
+	chara.add_baseATK(510);
 	chara.value[ER] += 0.459;
 	chara.value[EM] += 36 * 5;
 	chara.value[ATK] -= 0.05 * 5;
 	return chara;
 }
 
-// Ä§µ¼Ğ÷ÂÛ
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Magic_Guide(Tao chara)
+// é­”å¯¼ç»ªè®º
+auto Magic_Guide(derived_from_character auto chara)
 {
-	chara.baseATK += 354;
+	chara.add_baseATK(354);
 	chara.value[EM] += 187;
 	chara.value[bonus] += 0.24;
 	return chara;
 }
 
-// ÉñÀÖÖ®ÕæÒâ
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Kagura_s_Verity(Tao chara)
+// ç¥ä¹ä¹‹çœŸæ„
+auto Kagura_s_Verity(derived_from_character auto chara)
 {
-	chara.baseATK += 608;
+	chara.add_baseATK(608);
 	chara.value[CRD] += 0.662;
 	chara.e_bonus += 0.12 * 3;
 	chara.value[bonus] += 0.12;
 	return chara;
 }
 
-// ÊÔ×÷½ğçê
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Prototype_Amber(Tao chara)
+// è¯•ä½œé‡‘ç€
+auto Prototype_Amber(derived_from_character auto chara)
 {
-	chara.baseATK += 510;
+	chara.add_baseATK(510);
 	chara.value[HP] += 0.413;
 	return chara;
 }
 
-// ²»ÃğÔÂ»ª
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Everlasting_Moonglow(Tao chara)
+// ä¸ç­æœˆå
+auto Everlasting_Moonglow(derived_from_character auto chara)
 {
-	chara.baseATK += 608;
+	chara.add_baseATK(608);
 	chara.value[HP] += 0.496;
 	chara.value[healing] += 0.1;
 	return chara;
 }
 
-// ¼ÀÀñ²ĞÕÂ
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Sacrificial_Fragments(Tao chara)
+// ç¥­ç¤¼æ®‹ç« 
+auto Sacrificial_Fragments(derived_from_character auto chara)
 {
-	chara.baseATK += 454;
+	chara.add_baseATK(454);
 	chara.value[EM] += 221;
 	return chara;
 }
 
+// é—ç¥€ç‰ç‘
+auto Sacrificial_Jade(derived_from_character auto chara, bool stay_no_10s = true)
+{
+	chara.add_baseATK(454);
+	chara.value[CR] += .368;
+	if (stay_no_10s)
+	{
+		chara.value[HP] += .32;
+		chara.value[EM] += 40;
+	}
+	return chara;
+}
+
+// ä¸‡ä¸–æµæ¶Œå¤§å…¸
+auto Tome_of_the_Eternal_Flow(derived_from_character auto chara, int hp_change = 3)
+{
+	chara.add_baseATK(542);
+	chara.value[CRD] += .882;
+	chara.value[HP] += .16;
+	if (hp_change)
+	{
+		chara.value[bonus] += .14 * hp_change;
+	}
+	return chara;
+}
+
+// ä¸‡ä¸–æµæ¶Œå¤§å…¸
+auto Tome_of_the_Eternal_Flow_80(derived_from_character auto chara, int hp_change = 3)
+{
+	chara.add_baseATK(475);
+	chara.value[CRD] += .804;
+	chara.value[HP] += .16;
+	if (hp_change)
+	{
+		chara.value[bonus] += .14 * hp_change;
+	}
+	return chara;
+}