+++ Elemental_Resonance.h
@@ -13,36 +13,33 @@
 #include "Bloom.h"
 #include "Ayaka.h"
 #include "Gan_Yu.h"
+#include "Nuevillette.h"
 
 #include "buff.h"
 
-// 愈疗之水
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Soothing_Water(Tao chara)
+// 鎰堢枟涔嬫按
+auto Soothing_Water(derived_from_character auto chara)
 {
 	chara.value[HP] += 0.25;
 	return chara;
 }
 
-// 坚定之岩
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Enduring_Rock(Tao chara)
+// 鍧氬畾涔嬪博
+auto Enduring_Rock(derived_from_character auto chara)
 {
 	chara.value[bonus] += 0.15;
 	return chara;
 }
 
-// 热诚之火
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Fervent_Flames(Tao chara)
+// 鐑瘹涔嬬伀
+auto Fervent_Flames(derived_from_character auto chara)
 {
 	chara.value[ATK] += 0.25;
 	return chara;
 }
 
-// 粉碎之冰
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Shattering_Ice(Tao chara)
+// 绮夌涔嬪啺
+auto Shattering_Ice(derived_from_character auto chara)
 {
 	chara.value[CR] += 0.15;
 	return chara;