+++ Polearm.h
@@ -1,32 +1,29 @@
 #pragma once
 #include "random.h"
 
-// »¤Ä¦Ö®ÕÈ
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Staff_of_Homa(Tao chara)
+// æŠ¤æ‘©ä¹‹æ–
+auto Staff_of_Homa(derived_from_character auto chara)
 {
-	chara.baseATK += 608;
+	chara.add_baseATK(608);
 	chara.value[CRD] += 0.662;
 	chara.value[HP] += 0.2;
 	chara.HPrate += 0.018;
 	return chara;
 }
 
-// Ï»ÀïÃğ³½
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Dragon_s_Bane(Tao chara)
+// åŒ£é‡Œç­è¾°
+auto Dragon_s_Bane(derived_from_character auto chara)
 {
-	chara.baseATK += 454;
+	chara.add_baseATK(454);
 	chara.value[EM] += 221;
 	chara.value[bonus] += 0.36;
 	return chara;
 }
 
-// ºÍè±ğ°
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Primordial_Jade_Winged_Spear(Tao chara)
+// å’Œç’é¸¢
+auto Primordial_Jade_Winged_Spear(derived_from_character auto chara)
 {
-	chara.baseATK += 674;
+	chara.add_baseATK(674);
 	chara.value[CR] += 0.221;
 	chara.value[ATK] += 0.032 * 7;
 	chara.value[bonus] += 0.12;