+++ Sword.h
@@ -1,20 +1,18 @@
 #pragma once
 #include "random.h"
 
-// ÌìÄ¿Ó°´òµ¶
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Amenoma_Kageuchi(Tao chara)
+// å¤©ç›®å½±æ‰“åˆ€
+auto Amenoma_Kageuchi(derived_from_character auto chara)
 {
-	chara.baseATK += 454;
+	chara.add_baseATK(454);
 	chara.value[ATK] += 0.551;
 	return chara;
 }
 
-// ÎíÇÐÖ®»Ø¹â
-template<typename Tao, typename = enable_if_t<is_base_of<character, Tao>::value>>
-auto Mistsplitter_Reforged(Tao chara)
+// é›¾åˆ‡ä¹‹å›žå…‰
+auto Mistsplitter_Reforged(derived_from_character auto chara)
 {
-	chara.baseATK += 674;
+	chara.add_baseATK(674);
 	chara.value[CRD] += 0.441;
 	chara.value[bonus] += 0.12 + 0.28;
 	return chara;